requirements.txt
adal==1.2.7
altair==5.2.0
anyio==4.2.0
applicationinsights==0.11.10
argcomplete==3.2.1
attrs==23.1.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.1.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
contextlib2==21.6.0
cryptography==41.0.7
distro==1.8.0
docker==6.1.3
dotnetcore2==3.1.23
fusepy==3.0.1
gitdb==4.0.11
GitPython==3.1.40
holidays==0.39
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
isodate==0.6.1
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpickle==3.0.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kaggle==1.5.16
knack==0.11.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msal==1.26.0
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.23.5
oauthlib==3.2.2
opendatasets==0.1.22
packaging==23.2
pandas==1.5.3
paramiko==3.4.0
pathspec==0.12.1
Pillow==10.1.0
pkginfo==1.9.6
portalocker==2.8.2
protobuf==4.25.1
py4j==0.10.9.7
pyarrow==14.0.2
pyasn1==0.5.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
PySocks==1.7.1
pyspark==3.5.0
python-calendarific==1.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.15.2
scipy==1.11.4
SecretStorage==3.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
streamlit==1.29.0
streamlit-aggrid==0.3.4.post3
tabulate==0.9.0
tenacity==8.2.3
text-unidecode==1.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
webencodings==0.5.1
websocket-client==1.7.0
zipp==3.17.0
